change rowView by mockData

+++ LP-Besto/ViewModel/PlatViewModel.swift
@@ -22,7 +22,7 @@ class PlatViewModel: ObservableObject {
         plats.append(contentsOf: Plat.mockData)
     }
     
-    func addPlat(name: String, contenu: [Int], description: String, prix: Int) {
+    func addPlat(name: String, contenu: [Int], description: String, prix: Float) {
         let newPlat = Plat(name: name, contenu: contenu, description: description, prix: prix)
         plats.append(newPlat)
     }
